requirements.txt
pyyaml>=6.0
psutil>=5.9.0
requests>=2.31.0
python-dotenv>=1.0.0
colorama>=0.4.6
flask>=2.3.0
flask-cors>=4.0.0
flask-login>=0.6.2

